#184829218 Admin changes the status of delivered or cancelled-merged

+++ database/models/cart.js
@@ -19,7 +19,7 @@ module.exports = (sequelize, DataTypes) => {
       autoIncrement: true
     },
     productId: {
-      type: DataTypes.INTEGER(36),
+      type: DataTypes.INTEGER,
       allowNull: false
     },
     userId: {
